split up perf and non perf tests
